SimplePLTrainerを追加、関連クラスをエイリアスに変更
